extracted fixtures to fixtures file and moved to folder
